Removed mx-related tests from core.
